Factor createMutator out of Mutator.ts
