+++ src/createMutator.ts
@@ -1,0 +1,5 @@
+import Mutator from './Mutator';
+
+export default function createMutator<TState>(initialValue: TState) {
+    return new Mutator(initialValue);
+}
